# DevBoard.API/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Bu API'nin açıklaması",
        "title": "API Dökümanı",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Destek",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/user-roles": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-roles"
                ],
                "summary": "List all user roles",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.UserRole"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-roles"
                ],
                "summary": "Create a new user role assignment",
                "parameters": [
                    {
                        "description": "UserRole data",
                        "name": "userRole",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateUserRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.UserRole"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/user-roles/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-roles"
                ],
                "summary": "Get current user's roles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.UserRole"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/user-roles/role/{roleId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-roles"
                ],
                "summary": "Get user roles by role ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Role ID",
                        "name": "roleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.UserRole"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/user-roles/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-roles"
                ],
                "summary": "Get user role by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserRole ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.UserRole"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-roles"
                ],
                "summary": "Update user role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserRole ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "UserRole data",
                        "name": "userRole",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateUserRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.UserRole"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "user-roles"
                ],
                "summary": "Delete user role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserRole ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entities.Role": {
            "type": "object",
            "properties": {
                "createdBy": {
                    "type": "string"
                },
                "createdOn": {
                    "type": "string"
                },
                "deletedBy": {
                    "type": "string"
                },
                "deletedOn": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "lastModifiedBy": {
                    "type": "string"
                },
                "lastModifiedOn": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "userRoles": {
                    "description": "Relations",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.UserRole"
                    }
                }
            }
        },
        "entities.User": {
            "type": "object",
            "properties": {
                "createdBy": {
                    "type": "string"
                },
                "createdOn": {
                    "type": "string"
                },
                "deletedBy": {
                    "type": "string"
                },
                "deletedOn": {
                    "type": "string"
                },
                "deviceInfo": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "id": {
                    "description": "Props",
                    "type": "string"
                },
                "ipaddress": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "isEmailValidated": {
                    "type": "boolean"
                },
                "lastModifiedBy": {
                    "type": "string"
                },
                "lastModifiedOn": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "profilePictureUrl": {
                    "type": "string"
                },
                "refreshTokenExp": {
                    "type": "string"
                },
                "refreshTokenHash": {
                    "description": "Session properties",
                    "type": "string"
                },
                "userAgent": {
                    "type": "string"
                },
                "userRoles": {
                    "description": "Relations",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.UserRole"
                    }
                }
            }
        },
        "entities.UserRole": {
            "type": "object",
            "properties": {
                "createdBy": {
                    "type": "string"
                },
                "createdOn": {
                    "type": "string"
                },
                "deletedBy": {
                    "type": "string"
                },
                "deletedOn": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "lastModifiedBy": {
                    "type": "string"
                },
                "lastModifiedOn": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/entities.Role"
                },
                "roleID": {
                    "type": "string"
                },
                "user": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entities.User"
                        }
                    ]
                },
                "userID": {
                    "type": "string"
                }
            }
        },
...
        "handler.CreateUserRoleRequest": {
            "type": "object",
            "required": [
                "role_id"
            ],
            "properties": {
                "role_id": {
                    "type": "string"
                }
            }
        },
...
        "handler.UpdateUserRoleRequest": {
            "type": "object",
            "required": [
                "role_id",
                "user_id"
            ],
            "properties": {
                "role_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    }
}